evals/requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
attrs==24.3.0
autoevals==0.0.111
braintrust==0.0.177
braintrust-core==0.0.54
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chevron==0.14.0
deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
gitdb==4.0.11
gitpython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
humanloop==0.8.20
idna==3.10
importlib-metadata==8.4.0
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
levenshtein==0.26.1
openai==1.58.1
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-anthropic==0.33.9
opentelemetry-instrumentation-bedrock==0.33.9
opentelemetry-instrumentation-cohere==0.33.9
opentelemetry-instrumentation-groq==0.33.9
opentelemetry-instrumentation-openai==0.33.9
opentelemetry-instrumentation-replicate==0.33.9
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-semantic-conventions-ai==0.4.2
parse==1.20.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.3
pydantic-core==2.27.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-slugify==8.0.4
pyyaml==6.0.2
rapidfuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
setuptools==75.6.0
smmap==5.0.1
sniffio==1.3.1
sseclient-py==1.8.0
text-unidecode==1.3
tiktoken==0.8.0
tqdm==4.67.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.17.0
zipp==3.21.0
